requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
pyyaml>=6.0.0
